Bond flexion probabilities.
The new column is the probability that any given side chain bond of a binding residue will be flexed in real world ligand-bound proteins. The probability that the side chain will flex at all is this value times the number of rotatable bonds, so for example lysine has 4 rotatable side chain bonds so its probability of .0979 times four is .392 meaning almost 40% of all binding-site lysines are flexed in real proteins.