Going to another foolish social event.
So far lately I've gone to a funeral for someone I didn't like, several painful and boring social events that I didn't want to go to, met someone whose mealy mouth was so offensive I preemptively blocked
them, and now going to a birthday party for someone I wouldn't be friends with if given the choice. All of which time could have been better spent on PrimaryDock.